monitor/requirements.txt
twikit>=2.0.0
anthropic>=0.40.0
feedparser>=6.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
